[FIRRTL] Use symbols, not "dead wire" GCT Taps
Change Grand Central (GCT) Views and Taps to not add "dead wire" taps on
XMR'd operations marked with DontTouchAnnotations.  This prevents a
problem where the addition of these "dead wire" taps will block
deduplication if only one instance of a module is tapped (in
Chisel-emitted FIRRTL where all instances of a module are duplicated as
separate modules).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>